Register: use 64bit instead of 32 for Lamport's clock
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/concordant/sources/swiftcloud/java/SwiftCloud-branches/1pc@1176 049c2a90-bdf1-4f83-8191-3719d539f8e0

+++ src/swift/crdt/RegisterTxnLocal.java
@@ -9,10 +9,10 @@ import swift.crdt.operations.RegisterUpdate;
 
 public class RegisterTxnLocal<V extends Copyable> extends BaseCRDTTxnLocal<RegisterVersioned<V>> {
     private V val;
-    private int nextLamportClock;
+    private long nextLamportClock;
 
     public RegisterTxnLocal(CRDTIdentifier id, TxnHandle txn, CausalityClock clock, RegisterVersioned<V> creationState,
-            V val, int nextLamportClock) {
+            V val, long nextLamportClock) {
         super(id, txn, clock, creationState);
         this.val = val;
         this.nextLamportClock = nextLamportClock;
